probe: Cleanup the code
